api/pyproject.toml
[tool.poetry]
name = "cef assistant"
version = "0.1.0"
description = "CEF Assistant - Código de ética"
authors = ["Victor Resende <victoresende19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.8"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
gunicorn = "^23.0.0"
langchain-community = "^0.3.8"
pydantic = "^2.10.1"
python-dotenv = "^1.0.1"
pypdf = "^5.1.0"
openai = "^1.55.0"
faiss-cpu = "^1.9.0.post1"
tiktoken = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
